Rename TokenMaskService to CsrfTokenMask.
