+++ src/CsrfMiddleware.php
@@ -9,7 +9,7 @@ use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Server\MiddlewareInterface;
 use Psr\Http\Server\RequestHandlerInterface;
-use Yiisoft\Csrf\TokenMaskService\TokenMaskServiceInterface;
+use Yiisoft\Csrf\TokenMask\CsrfTokenMaskInterface;
 use Yiisoft\Csrf\TokenStorage\CsrfTokenStorageInterface;
 use Yiisoft\Http\Method;
 use Yiisoft\Http\Status;
@@ -26,16 +26,16 @@ final class CsrfMiddleware implements MiddlewareInterface
 
     private ResponseFactoryInterface $responseFactory;
     private CsrfTokenStorageInterface $storage;
-    private TokenMaskServiceInterface $tokenMaskService;
+    private CsrfTokenMaskInterface $tokenMask;
 
     public function __construct(
         ResponseFactoryInterface $responseFactory,
         CsrfTokenStorageInterface $storage,
-        TokenMaskServiceInterface $tokenMaskService
+        CsrfTokenMaskInterface $tokenMask
     ) {
         $this->responseFactory = $responseFactory;
         $this->storage = $storage;
-        $this->tokenMaskService = $tokenMaskService;
+        $this->tokenMask = $tokenMask;
     }
 
     public function process(ServerRequestInterface $request, RequestHandlerInterface $handler): ResponseInterface
@@ -101,6 +101,6 @@ final class CsrfMiddleware implements MiddlewareInterface
             $token = \reset($headers);
         }
 
-        return is_string($token) ? $this->tokenMaskService->remove($token) : null;
+        return is_string($token) ? $this->tokenMask->remove($token) : null;
     }
 }
