+++ src/CsrfToken.php
@@ -5,21 +5,21 @@ declare(strict_types=1);
 namespace Yiisoft\Csrf;
 
 use LogicException;
-use Yiisoft\Csrf\TokenMaskService\TokenMaskServiceInterface;
+use Yiisoft\Csrf\TokenMask\CsrfTokenMaskInterface;
 use Yiisoft\Csrf\TokenStorage\CsrfTokenStorageInterface;
 
 final class CsrfToken
 {
 
     private CsrfTokenStorageInterface $storage;
-    private TokenMaskServiceInterface $tokenMaskService;
+    private CsrfTokenMaskInterface $tokenMask;
 
     public function __construct(
         CsrfTokenStorageInterface $storage,
-        TokenMaskServiceInterface $tokenMaskService
+        CsrfTokenMaskInterface $tokenMask
     ) {
         $this->storage = $storage;
-        $this->tokenMaskService = $tokenMaskService;
+        $this->tokenMask = $tokenMask;
     }
 
     /**
@@ -32,6 +32,6 @@ final class CsrfToken
         if (empty($token)) {
             throw new LogicException('CSRF token is not defined.');
         }
-        return $this->tokenMaskService->apply($token);
+        return $this->tokenMask->apply($token);
     }
 }
